Fix constexpr example
In the example it's not guaranteed that `int a = f(5)` will be
computed at compile time (although that will be the case with most
if not all compilers), but we can force it by adding a `constexpr`.

This also serves to show that variables can be constexpr, not just
functions.