Avoid creating shared library in code/hello
Several students were confused by that extra file and/or forgot to
`export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:.` and/or asked what that
magic spell did.

Explanation of the compilation chain is on the last day, we could
leave creation of shared objects and LD_LIBRARY_PATH for that day.