New exceptions section, recommended modern use
New exceptions section with more illustrative examples
of recommended use. This includes some discussion of
what one should use exceptions for, what one should not
use them for. This includes some performance considerations.

All of the deprecated throw() syntax is removed, and
the slide on the more current nothrow is expanded.

A few additional LaTeX ephemeral files are added to
.gitignore (arise when processing with MiKTeX).