[concurrency] Work on mutex
- Remove global variables from mutex example, use lambdas.

In order to not show using globals, I thought using lambdas to capture
local variables is cleaner.
I moved to C++17, since we say in the lecture now that people should
alwyas use scoped_lock.

I also return an exit status that captures whether or not there was a
race.