Updated/modernised the CMake configuration of all code examples.
Made sure that all examples would declare/use the "solution" target
consistently.

Made sure that the solutions (and the main sources where appropriate)
would compile warning-free with MSVC.

Removed the "solution" for the constness example, as it was the same
broken source file that we give to the students to fix up/study.

Reduced the C++ standard used by the "lambdas" exercise to C++11, as
it is not compatible with C++17 or newer, and it prints a warning
about using deprecated code with C++14.

Made some tiny text updates in asan/README.md.