Add \usepackage{morewrites} to overcome "no room for new \write" error
This error happens when the user or the included packages try to
allocate too many streams for auxiliary files. It seems to be caused
by the lists of good practices, exercises, and godbolt links in this
document (possibly in combination with the already included packages).

See https://ctan.org/pkg/morewrites for more information.