Simplify debug exercise.
The previous version had a lot of code that took time to process. Here,
we replace the random generator and shuffling of the vector by a
reverse operation. In addition, we can now check that the result is correct
using std::is_sorted.

Also run clang-format with a new style file that will be merged in a future PR.

Co-authored-by: Philippe Canal <pcanal@fnal.gov>