Remove <=> and hidden friends from operators exercise.
This exercise is used both in the essentials and the advanced course. We
can therefore not require students to work with hidden friends and <=>.
Here, all mentions of <=> are removed (there is a dedicated exercise on
the third day of the advanced course), and the tasks are rephrased such
that students can work on this exercise also in the essentials course.

The diff between the original file and the solution is minimised, so
viewing the solution with a diff program hopefully helps the students.